+++ CHANGELOG.md
@@ -14,6 +14,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Updated `@janiscommerce/api` to `6.x.x`
 
+## [4.1.1] - 2020-08-10
+### Fixed
+- Max page size is validated as 100 records
+
+## [4.1.0] - 2020-07-29
+### Added
+- `formatFilters` method
+
 ## [4.0.1] - 2020-06-23
 ### Fixed
 - Filters with an object as value are now valid